Set onboard location per network
Allow location to be set on a per network basis. This allows
hotspots that support both networks to set location on IOT
and opt to not set the location on MOBILE.
