+++ package.json
@@ -63,7 +63,7 @@
     "@types/bn.js": "^5.1.1",
     "@types/jest": "^26.0.0",
     "@types/lodash": "^4.14.172",
-    "@types/react": "^16.9.19",
+    "@types/react": "^18.0.24",
     "@types/react-native": "^0.64.14",
     "base-64": "^1.0.0",
     "buffer": "^6.0.3",
@@ -78,7 +78,7 @@
     "pod-install": "^0.1.0",
     "prettier": "^2.0.5",
     "protobufjs": "^6.11.2",
-    "react": "16.13.1",
+    "react": "18.2.0",
     "react-native": "0.63.4",
     "react-native-ble-plx": "^2.0.2",
     "react-native-builder-bob": "^0.18.1",
